PR2: extract studio_core.services.actions + TUI shims
Move the unified ActionSpec registry (launcher + cockpit) into
studio_core so TUI/Web/API share one safe argv/form catalog.
tools/cli/tui/actions.py re-exports for backward compatibility.
